+++ src/state/gameday.rs
@@ -1,0 +1,24 @@
+/// Store which panels should be rendered in the Gameday tab.
+#[derive(Debug, Copy, Clone)]
+pub struct GamedayPanels {
+    pub info: bool,
+    pub at_bat: bool,
+    pub boxscore: bool,
+}
+
+impl GamedayPanels {
+    /// Return the number of panels that are active.
+    pub fn count(&self) -> usize {
+        self.info as usize + self.at_bat as usize + self.boxscore as usize
+    }
+}
+
+impl Default for GamedayPanels {
+    fn default() -> Self {
+        GamedayPanels {
+            info: true,
+            at_bat: true,
+            boxscore: false,
+        }
+    }
+}
