+++ src/state/refresher.rs
@@ -1,5 +1,5 @@
+use crate::NetworkRequest;
 use crate::app::{App, MenuItem};
-use crate::messages::NetworkRequest;
 use std::time::Duration;
 use tokio::sync::mpsc;
 use tokio::time::interval;
